read write tested.

+++ ezlopi-cloud/settings/settings.c
@@ -26,10 +26,10 @@ void settings_list(cJSON *cj_request, cJSON *cj_response) {
             s_ezlopi_settings_t *settings_list = ezlopi_settings_get_settings_list();
 
             // Calculate the number of settings
-            size_t num_settings = ezlopi_settings_get_settings_count();
+            uint16_t num_settings = ezlopi_settings_get_settings_count();
 
             // Add each setting to the "settings" array
-            for (size_t i = 0; i < num_settings; i++) 
+            for (uint16_t i = 0; i < num_settings; i++) 
             {
                 cJSON *cj_setting = cJSON_CreateObject();
                 cJSON_AddItemToArray(cj_settings, cj_setting);
@@ -49,7 +49,7 @@ void settings_list(cJSON *cj_request, cJSON *cj_response) {
                         cJSON_AddItemToObject(cj_setting, "enum", cj_enum);
 
                         // Add enum values to the "enum" array
-                        for (size_t j = 0; j < EZLOPI_SETTINGS_MAX_ENUM_VALUES; j++) {
+                        for (uint16_t j = 0; j < EZLOPI_SETTINGS_MAX_ENUM_VALUES; j++) {
                             const char *enum_value = settings_list[i].enum_values[j];
                             if (enum_value != NULL) {
                                 cJSON_AddItemToArray(cj_enum, cJSON_CreateString(enum_value));
@@ -57,7 +57,6 @@ void settings_list(cJSON *cj_request, cJSON *cj_response) {
                                 break;
                             }
                         }                        
-
                         break;
                     case EZLOPI_SETTINGS_TYPE_BOOL:
                         cJSON_AddBoolToObject(cj_setting, "value", settings_list[i].value.bool_value);
